separate and rename fsm_core. use the fsm configuration handler
